refactor autoscaling_is_paused method
