+++ tests/cli/toggle_cli_test.py
